requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
sentence-transformers
faiss-cpu
python-dotenv
pypdf
requests
